Bump PumpX2 to v1.8.8; use FSL getValue(); fix SetMaxBasalLimitRequest
- pumpx2_version 1.8.8 (JitPack).
- VicoCgmChart: read FSL2/FSL3 glucose via parsed.value after parse(); drop cargo helper.
- SetMaxBasalLimitRequest now takes long milliunits (PumpX2 API change).
- Remove obsolete pumpx2 FSL patch; document 1.8.8 in patches/README.
- FSL3 unit test: 26-byte cargo with LE 149 at offset 16 per PumpX2 parse().

+++ mobile/src/main/java/com/jwoglom/controlx2/presentation/screens/sections/Debug.kt
@@ -2152,7 +2152,7 @@ fun triggerSetMaxBasalLimitRequestMessage(
     input.inputType = InputType.TYPE_CLASS_NUMBER
     builder.setView(input)
     builder.setPositiveButton("Send") { dialog, _ ->
-        sendPumpCommands(SendType.DEBUG_PROMPT, listOf(SetMaxBasalLimitRequest(input.text.toString().toInt())))
+        sendPumpCommands(SendType.DEBUG_PROMPT, listOf(SetMaxBasalLimitRequest(input.text.toString().toLong())))
         dialog.dismiss()
     }
     builder.setNegativeButton("Cancel") { dialog, _ -> dialog.cancel() }
